tx-verifier: Move sig destination getter to test code
